refactor(api): extract BaseRagEmbeddingHelper from rag helpers
