+++ packages/api/src/helper/lib/base-rag-helper.ts
@@ -12,12 +12,6 @@ import { HelperName, HelperType } from '../types';
 
 import BaseHelper from './base-helper';
 
-/** One window of content text, ready to be embedded and stored. */
-export interface RagChunk {
-  index: number;
-  text: string;
-}
-
 /**
  * Base class for Retrieval-Augmented Generation (RAG) helpers.
  *
@@ -66,72 +60,6 @@ export abstract class BaseRagHelper<
    * Optional: reconcile or rebuild the helper from the source database.
    */
   reindex?(): Promise<void>;
-
-  /**
-   * Splits canonical content text into deterministic, overlapping character
-   * windows. Newline boundaries are preferred in the latter half of a window;
-   * oversized lines are hard-split at the configured size.
-   *
-   * Lives here so every embedding-backed helper chunks identically: the window
-   * boundaries are baked into each stored vector, so two helpers that disagreed
-   * about them would produce indexes that cannot be compared or migrated
-   * between. Determinism matters for the same reason — re-chunking unchanged
-   * text must yield byte-identical windows, or every reconciliation pass would
-   * consider the whole corpus stale.
-   *
-   * @throws RangeError when the chunk size or overlap is not usable.
-   */
-  protected chunkSearchText(
-    source: string,
-    chunkSize: number,
-    overlap: number,
-  ): RagChunk[] {
-    if (!Number.isInteger(chunkSize) || chunkSize <= 0) {
-      throw new RangeError('Chunk size must be a positive integer.');
-    }
-    if (!Number.isInteger(overlap) || overlap < 0 || overlap >= chunkSize) {
-      throw new RangeError(
-        'Chunk overlap must be a non-negative integer smaller than chunk size.',
-      );
-    }
-
-    const text = source.replace(/\r\n?/g, '\n');
-    const chunks: RagChunk[] = [];
-    let start = 0;
-
-    while (start < text.length) {
-      const hardEnd = Math.min(start + chunkSize, text.length);
-      let end = hardEnd;
-
-      if (hardEnd < text.length) {
-        const minimumBoundary = start + Math.floor(chunkSize / 2);
-        const paragraphBoundary = text.lastIndexOf('\n\n', hardEnd - 1);
-        const lineBoundary = text.lastIndexOf('\n', hardEnd - 1);
-        if (paragraphBoundary >= minimumBoundary) {
-          end = paragraphBoundary + 2;
-        } else if (lineBoundary >= minimumBoundary) {
-          end = lineBoundary + 1;
-        }
-      }
-
-      const value = text.slice(start, end).trim();
-      if (value) {
-        chunks.push({
-          index: chunks.length,
-          text: value,
-        });
-      }
-
-      if (end >= text.length) {
-        break;
-      }
-
-      const nextStart = Math.max(0, end - overlap);
-      start = nextStart > start ? nextStart : end;
-    }
-
-    return chunks;
-  }
 }
 
 export default BaseRagHelper;
